# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Prometheus-X Consent Manager",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://{{apiurl}}"
    }
  ],
...
  "paths": {
...
    "/users/login": {
      "post": {
        "tags": ["users"],
        "summary": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "email": "<string>",
                  "password": "<string>"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": "853"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"355-HC4VvlHixu7/doolM3v5yr8EkS8\""
                }
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "example": "consentmanagersessid=s%3A73db37f0-ccc9-49e6-8f0f-32b3c17f55d1.I4oF6QQBTkAkOjbr2d9f3tov%2F8kNZk7%2FqbSjXz8p1gI; Path=/; Expires=Sat, 02 Sep 2023 15:00:35 GMT; HttpOnly; SameSite=Lax"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Fri, 01 Sep 2023 15:00:35 GMT"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "user": {
                    "_id": "64f1fc8fbd8a3583c76d62bc",
                    "firstName": "pblt8tki",
                    "lastName": "RWFdeY0P",
                    "email": "KmWIBNF7mf@example.com",
                    "identifiers": [],
                    "jsonld": "{\n  \"@context\": \"http://schema.org\",\n  \"@type\": \"Person\",\n  \"name\": \"pblt8tki RWFdeY0P\",\n  \"email\": \"KmWIBNF7mf@example.com\",\n  \"url\": \"undefined:3000/v1/users/64f1fc8fbd8a3583c76d62bc\"\n}",
                    "schema_version": "v0.1.0",
                    "createdAt": "2023-09-01T15:00:31.044Z",
                    "updatedAt": "2023-09-01T15:00:35.883Z",
                    "__v": 0
                  },
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2NGYxZmM4ZmJkOGEzNTgzYzc2ZDYyYmMiLCJlbWFpbCI6IkttV0lCTkY3bWZAZXhhbXBsZS5jb20iLCJzY29wZXMiOlsiUmVhZCB1c2VyIGRhdGEiLCJNb2RpZnkgdXNlciBkYXRhIl0sImlhdCI6MTY5MzU4MDQzNSwiZXhwIjoxNjkzNTg0MDM1fQ.x74CzmjjHNY4CTpaOvH44P2rRfB7wjSSMJLPyNkHZjA",
                  "refreshToken": "c8197010c01bdf601a6bfdb72cf46bf98c3388fcf04771f87bee96f8c436bf77"
                }
              }
            }
          }
        }
      }
    },
...
  }
}